+++ modules/mesh-layers/src/scenegraph-layer/scenegraph-layer-fragment.glsl.js
@@ -1,17 +1,39 @@
 export default `\
-#ifdef HAS_UV
-  varying vec2 vTEXCOORD_0;
-  uniform sampler2D u_BaseColorSampler;
+#if (__VERSION__ < 300)
+  #define _varying varying
+  #define _texture2D texture2D
+  #define fragmentColor gl_FragColor
+#else
+  #define _varying in
+  #define _texture2D texture
+  out vec4 fragmentColor;
 #endif
-varying vec4 vColor;
+
+// Uniforms
+// none
+
+// Varying
+// MODULE_PBR contains all the varying definitions needed
+#ifndef MODULE_PBR
+  _varying vec4 vColor;
+
+  #if defined(HAS_UV) && defined(HAS_BASECOLORMAP)
+    _varying vec2 vTEXCOORD_0;
+    uniform sampler2D u_BaseColorSampler;
+  #endif
+#endif
 
 void main(void) {
-  #ifdef HAS_UV
-    gl_FragColor = (vColor / 255.) * texture2D(u_BaseColorSampler, vTEXCOORD_0);
+  #ifdef MODULE_PBR
+    fragmentColor = pbr_filterColor(vec4(0));
   #else
-    gl_FragColor = vColor / 255.;
+    #if defined(HAS_UV) && defined(HAS_BASECOLORMAP)
+      fragmentColor = vColor * _texture2D(u_BaseColorSampler, vTEXCOORD_0);
+    #else
+      fragmentColor = vColor;
+    #endif
   #endif
 
-  gl_FragColor = picking_filterPickingColor(gl_FragColor);
+  fragmentColor = picking_filterPickingColor(fragmentColor);
 }
 `;